GDALMDArray::GetResampled(): make GLT orthorectification available to all drivers, including virtual group returned by GDALGroup::SubsetDimensionFromSelection()
